Revert "Convert Dashboard to Server-Side Rendering with runWithAmplifyServerContext"
